Simplify by deleting abstract safe sql builder and removing SafeSqlAppender
