+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/Alias.java
@@ -2,7 +2,7 @@ package org.dhatim.safesql.builder;
 
 import java.util.Objects;
 import org.dhatim.safesql.SafeSql;
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 import org.dhatim.safesql.SafeSqlUtils;
 import org.dhatim.safesql.SafeSqlizable;
 
@@ -50,9 +50,9 @@ public final class Alias implements SafeSqlizable {
     public SafeSql toSafeSql() {
         return SafeSqlUtils.fromIdentifier(name);
     }
-    
+
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         builder.identifier(name);
     }
