+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/CommonTableExpression.java
@@ -3,7 +3,7 @@ package org.dhatim.safesql.builder;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 import org.dhatim.safesql.SafeSqlizable;
 
 public class CommonTableExpression implements SafeSqlizable {
@@ -15,19 +15,19 @@ public class CommonTableExpression implements SafeSqlizable {
     public CommonTableExpression(String name, SqlQuery query) {
         this(name, Collections.emptyList(), query);
     }
-    
+
     public CommonTableExpression(String name, List<String> columnNames, SqlQuery query) {
         this.name = name;
         this.query = query;
         this.columnNames = new ArrayList<>(columnNames);
     }
-    
+
     public String getName() {
         return name;
     }
-    
+
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         builder.identifier(name);
         if (!columnNames.isEmpty()) {
             builder.joinedSqlizables(", ", "(", ")", columnNames.stream().map(Identifier::new));
@@ -36,5 +36,5 @@ public class CommonTableExpression implements SafeSqlizable {
                 .append(query)
                 .append(")");
     }
-    
+
 }
