+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/Compute.java
@@ -1,6 +1,6 @@
 package org.dhatim.safesql.builder;
 
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 
 public class Compute implements Operand {
 
@@ -13,21 +13,21 @@ public class Compute implements Operand {
         this.operator = op;
         this.right = operand2;
     }
-    
+
     public Operand getLeftOperand() {
         return left;
     }
-    
+
     public Operand getRightOperand() {
         return right;
     }
-    
+
     public MathOperator getOperator() {
         return operator;
     }
-    
+
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         builder.append("(")
             .append(left)
             .append(" ")
@@ -36,11 +36,11 @@ public class Compute implements Operand {
             .append(right)
             .append(")");
     }
-    
+
     public static Compute add(Operand op1, Operand op2) {
         return new Compute(op1, MathOperator.ADD, op2);
     }
-    
+
     public static Compute sub(Operand op1, Operand op2) {
         return new Compute(op1, MathOperator.SUB, op2);
     }
