+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/Concat.java
@@ -3,11 +3,10 @@ package org.dhatim.safesql.builder;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 
 public class Concat implements Operand {
-    
+
     private final ArrayList<Operand> list = new ArrayList<>();
 
     public Concat(Collection<? extends Operand> c) {
@@ -16,19 +15,19 @@ public class Concat implements Operand {
         }
         list.addAll(c);
     }
-    
+
     public Concat(Operand left, Operand right) {
         list.add(left);
         list.add(right);
     }
-    
+
     public Concat(Operand left, Operand right, Operand... others) {
         this(left, right);
         list.addAll(Arrays.asList(others));
     }
-    
+
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         builder.joinedSqlizables(" || ", list);
     }
