+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/From.java
@@ -3,50 +3,50 @@ package org.dhatim.safesql.builder;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 import org.dhatim.safesql.SafeSqlizable;
 
 public abstract class From extends AbstractHasJointure implements SafeSqlizable {
-    
+
     private static class TableFrom extends From {
-        
+
         private final String schema;
         private final String tableName;
-        
+
         private TableFrom(String schema, String tableName, Alias alias, List<String> columnAliases) {
             super(alias, columnAliases);
             this.schema = schema;
             this.tableName = tableName;
         }
-        
+
         @Override
-        protected void render(SafeSqlAppendable sb) {
+        protected void render(SafeSqlBuilder sb) {
             if (schema != null) {
                 sb.identifier(schema).append(".");
             }
             sb.identifier(tableName);
         }
-        
+
     }
-    
+
     private static class SubQueryFrom extends From {
-        
+
         private final SqlQuery query;
 
         private SubQueryFrom(SqlQuery query, Alias alias, List<String> columnAliases) {
             super(alias, columnAliases);
             this.query = query;
         }
-        
+
         @Override
-        protected void render(SafeSqlAppendable sb) {
+        protected void render(SafeSqlBuilder sb) {
             sb.append('(');
             sb.append(query);
             sb.append(')');
         }
-        
+
     }
-    
+
     private final Alias alias;
     private final List<String> columnAliases;
 
@@ -59,10 +59,10 @@ public abstract class From extends AbstractHasJointure implements SafeSqlizable 
         return alias;
     }
 
-    protected abstract void render(SafeSqlAppendable sb);
+    protected abstract void render(SafeSqlBuilder sb);
 
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         render(builder);
         if (alias != null) {
             builder.append(" ").append(alias);
@@ -75,19 +75,19 @@ public abstract class From extends AbstractHasJointure implements SafeSqlizable 
             builder.append(" ").joinedSqlizables(" ", jointures);
         }
     }
-    
+
     public static From table(String schema, String tableName, Alias alias) {
         return new TableFrom(schema, tableName, alias, Collections.emptyList());
     }
-    
+
     public static From table(String schema, String tableName, Alias alias, List<String> columnAliases) {
         return new TableFrom(schema, tableName, alias, columnAliases);
     }
-    
+
     public static From query(SqlQuery query, Alias alias) {
         return new SubQueryFrom(query, alias, Collections.emptyList());
     }
-    
+
     public static From query(SqlQuery query, Alias alias, List<String> columnAliases) {
         return new SubQueryFrom(query, alias, columnAliases);
     }
