+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/Literal.java
@@ -1,17 +1,17 @@
 package org.dhatim.safesql.builder;
 
 import org.dhatim.safesql.SafeSql;
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 import org.dhatim.safesql.SafeSqlUtils;
 
 public class Literal implements Operand {
-    
+
     private static final Literal NULL_LITERAL = new Literal(null) {
         @Override
-        public void appendTo(SafeSqlAppendable builder) {
+        public void appendTo(SafeSqlBuilder builder) {
             builder.append("NULL");
         }
-        
+
         @Override
         public SafeSql toSafeSql() {
             return SafeSqlUtils.fromConstant("NULL");
@@ -23,19 +23,19 @@ public class Literal implements Operand {
     protected Literal(String value) {
         this.value = value;
     }
-    
+
     @Deprecated
     public static Literal from(String value) {
         return of(value);
     }
-    
+
     public static Literal of(String value) {
         return value == null ? NULL_LITERAL : new Literal(value);
     }
-    
+
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         builder.literal(value);
     }
-    
+
 }
