+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/MathOperator.java
@@ -1,7 +1,7 @@
 package org.dhatim.safesql.builder;
 
 import org.dhatim.safesql.SafeSql;
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 import org.dhatim.safesql.SafeSqlUtils;
 
 public enum MathOperator implements Operator {
@@ -10,13 +10,13 @@ public enum MathOperator implements Operator {
     MUL("*"),
     DIV("/"),
     MOD("%");
-    
+
     private final String sql;
 
     private MathOperator(String sql) {
         this.sql = sql;
     }
-    
+
     @Override
     public String toString() {
         return sql;
@@ -26,10 +26,10 @@ public enum MathOperator implements Operator {
     public SafeSql toSafeSql() {
         return SafeSqlUtils.fromConstant(sql);
     }
-    
+
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         builder.append(sql);
     }
-    
+
 }
