+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/NamedWindow.java
@@ -1,37 +1,37 @@
 package org.dhatim.safesql.builder;
 
 import java.util.List;
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 
 public class NamedWindow extends Window {
 
     private final String name;
-    
+
     public NamedWindow(String name) {
         this.name = name;
     }
-    
+
     public NamedWindow(String name, List<Operand> partition) {
         super(partition);
         this.name = name;
     }
-    
+
     public String getName() {
         return name;
     }
-    
+
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         builder.identifier(name)
             .append(" AS ");
         super.appendTo(builder);
     }
-    
+
     @Override
     public int hashCode() {
         return name.hashCode();
     }
-    
+
     @Override
     public boolean equals(Object obj) {
         if (obj == null) {
@@ -45,5 +45,5 @@ public class NamedWindow extends Window {
         }
         return false;
     }
-    
+
 }
