+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/Order.java
@@ -1,18 +1,18 @@
 package org.dhatim.safesql.builder;
 
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 import org.dhatim.safesql.SafeSqlizable;
 
 public enum Order implements SafeSqlizable {
     ASC {
         @Override
-        public void appendTo(SafeSqlAppendable builder) {
+        public void appendTo(SafeSqlBuilder builder) {
             builder.append("ASC");
         }
     },
     DESC {
         @Override
-        public void appendTo(SafeSqlAppendable builder) {
+        public void appendTo(SafeSqlBuilder builder) {
             builder.append("DESC");
         }
     };
