+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/Over.java
@@ -1,27 +1,27 @@
 package org.dhatim.safesql.builder;
 
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 
 public class Over implements Operand {
 
     private final Call windowFunction;
     private final Window window;
-    
+
     public Over(Call windowFunction, Window window) {
         this.windowFunction = windowFunction;
         this.window = window;
     }
-    
+
     public Window getWindow() {
         return window;
     }
-    
+
     public Call getWindowFunction() {
         return windowFunction;
     }
-    
+
     @Override
-    public void appendTo(SafeSqlAppendable sb) {
+    public void appendTo(SafeSqlBuilder sb) {
         sb.append(windowFunction);
         sb.append(" OVER ");
         if (window instanceof NamedWindow) {
