+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/SafeSqlQuery.java
@@ -1,10 +1,10 @@
 package org.dhatim.safesql.builder;
 
 import org.dhatim.safesql.SafeSql;
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 
 public class SafeSqlQuery implements SqlQuery {
-    
+
     private final SafeSql sql;
 
     public SafeSqlQuery(SafeSql sql) {
@@ -12,7 +12,7 @@ public class SafeSqlQuery implements SqlQuery {
     }
 
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         builder.append(sql);
     }
