+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/UnaryOperator.java
@@ -1,20 +1,20 @@
 package org.dhatim.safesql.builder;
 
 import org.dhatim.safesql.SafeSql;
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 import org.dhatim.safesql.SafeSqlUtils;
 import org.dhatim.safesql.SafeSqlizable;
 
 public enum UnaryOperator implements SafeSqlizable {
     POSITIVE("+"),
     NEGATIVE("-");
-    
+
     private final String sql;
 
     private UnaryOperator(String sql) {
         this.sql = sql;
     }
-    
+
     @Override
     public String toString() {
         return sql;
@@ -24,9 +24,9 @@ public enum UnaryOperator implements SafeSqlizable {
     public SafeSql toSafeSql() {
         return SafeSqlUtils.fromConstant(sql);
     }
-    
+
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         builder.append(sql);
     }
 }
