+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/Value.java
@@ -1,7 +1,7 @@
 package org.dhatim.safesql.builder;
 
 import org.dhatim.safesql.SafeSql;
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 import org.dhatim.safesql.SafeSqlUtils;
 
 public class Value implements Operand {
@@ -16,21 +16,21 @@ public class Value implements Operand {
     public SafeSql toSafeSql() {
         return SafeSqlUtils.escape(data);
     }
-    
+
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         builder.param(data);
     }
 
     public Object value() {
         return data;
     }
-    
+
     @Override
     public String toString() {
         return "Value{" + data.getClass().getSimpleName() + ":" + data + "}";
     }
-    
+
     public static Value of(Object data) {
         if (data instanceof Boolean) {
             return new BooleanValue((Boolean) data);
