+++ safesql-builder/src/main/java/org/dhatim/safesql/builder/Window.java
@@ -2,38 +2,38 @@ package org.dhatim.safesql.builder;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.dhatim.safesql.SafeSqlAppendable;
+import org.dhatim.safesql.SafeSqlBuilder;
 import org.dhatim.safesql.SafeSqlizable;
 
 public class Window implements SafeSqlizable {
 
     private final List<Operand> partition = new ArrayList<>();
-    
+
     protected Window() {
     }
-    
+
     public Window(List<Operand> partition) {
         this.partition.addAll(partition);
     }
-    
+
     public List<Operand> getPartition() {
         return new ArrayList<>(partition);
     }
-    
+
     public void setPartition(List<Operand> newPartitions) {
         partition.clear();
         partition.addAll(newPartitions);
     }
-    
+
     public void addPartition(Operand operand) {
         partition.add(operand);
     }
 
     @Override
-    public void appendTo(SafeSqlAppendable builder) {
+    public void appendTo(SafeSqlBuilder builder) {
         builder.append("(PARTITION BY ")
                 .joinedSqlizables(", ", partition)
                 .append(')');
     }
-    
+
 }
