+++ safesql/src/main/java/org/dhatim/safesql/SafeSqlizable.java
@@ -2,7 +2,7 @@ package org.dhatim.safesql;
 
 @FunctionalInterface
 public interface SafeSqlizable {
-    
+
     /**
      * Returns a {@link SafeSql} version of this object
      * @return a {@link SafeSql}
@@ -12,6 +12,6 @@ public interface SafeSqlizable {
         appendTo(sb);
         return sb.toSafeSql();
     }
-    
-    void appendTo(SafeSqlAppendable builder);
+
+    void appendTo(SafeSqlBuilder builder);
 }
